Fix warnings in utilities

+++ src/utils/xrLC/xrSaveOGF.cpp
@@ -22,7 +22,7 @@ u16		RegisterShader		(LPCSTR T)
 	for (u32 it=0; it<g_Shaders.size(); it++)
 		if (0==_stricmp(T,g_Shaders[it]))	return it;
 	g_Shaders.push_back		(xr_strdup(T));
-	return g_Shaders.size	()-1;
+	return (u16) g_Shaders.size	()-1;
 }
 
 void	geom_batch_average	(u32 verts, u32 faces)
